Move into it's own derivation file.
